Buy With Fiat API // Scene_Pay

+++ Assets/Thirdweb/Core/Scripts/Pay/Types.GetBuyWithCryptoQuote.cs
@@ -34,7 +34,7 @@ namespace Thirdweb.Pay
         public string ToAmountWei { get; set; }
 
         [JsonProperty("maxSlippageBPS")]
-        public int? MaxSlippageBPS { get; set; }
+        public double? MaxSlippageBPS { get; set; }
 
         public BuyWithCryptoQuoteParams(
             string fromAddress,
@@ -46,7 +46,7 @@ namespace Thirdweb.Pay
             BigInteger? toChainId = null,
             string toAmount = null,
             string toAmountWei = null,
-            int? maxSlippageBPS = null
+            double? maxSlippageBPS = null
         )
         {
             FromAddress = fromAddress;
@@ -182,7 +182,7 @@ namespace Thirdweb.Pay
         public Estimated Estimated { get; set; }
 
         [JsonProperty("maxSlippageBPS")]
-        public int MaxSlippageBPS { get; set; }
+        public double MaxSlippageBPS { get; set; }
 
         [JsonProperty("bridge")]
         public string Bridge { get; set; }
